Rearranging files and folders to be consistent for a proper python code
